+++ attributes/default.rb
@@ -1,6 +1,6 @@
 default['archiva']['mirror']   = 'http://archive.apache.org/dist/archiva/'
-default['archiva']['version']  = '2.1.1'
-default['archiva']['checksum'] = 'e1c3245b73f8f6aceae05527331cd1902dd5edb647056e2f83269fc0e2e8ace7'
+default['archiva']['version']  = '2.2.4'
+default['archiva']['checksum'] = 'ad099c451529d566db8b5ad6cf2d8bafc6aa2cc3d225c39fd374cad420a6643c'
 
 default['archiva']['install_path'] = '/opt/'
 default['archiva']['home']         = '/opt/archiva'
@@ -11,8 +11,13 @@ default['archiva']['web_port']     = '8080'
 default['archiva']['web_domain']   = 'archiva.example.com'
 default['archiva']['web_template'] = 'default'
 
-default['archiva']['nginx']      = 'default'
 default['archiva']['nginx_uri']  = '/'
 default['archiva']['nginx_port'] = '80'
 
-set['java']['jdk_version'] = '7'
+default['java']['jdk_version'] =
+  value_for_platform(
+    'debian' => {
+      '>= 10.0' => '11',
+    },
+    'default' => '8'
+  )