+++ recipes/default.rb
@@ -1,5 +1,5 @@
 #
-# Cookbook Name:: archiva
+# Cookbook:: archiva
 # Recipe:: default
 # Author:: Jorge Espada <espada.jorge@gmail.com>
 #
@@ -33,13 +33,19 @@ ark 'archiva' do
   action      :install
 end
 
+# We supply our own, slightly modified service script because the
+# one supplied by the package, is not compatible with SystemD.
+# This is a copy that script, modified to be compatible.
+# Credit to @torjeh for this contribution in #14
+cookbook_file "#{node['archiva']['home']}/bin/archiva"
+
 link '/etc/init.d/archiva' do
   to "#{node['archiva']['home']}/bin/archiva"
 end
 
 [
   'bin/wrapper-linux-x86-32',
-  'lib/libwrapper-linux-x86-32.so'
+  'lib/libwrapper-linux-x86-32.so',
 ].each do |file|
   file "#{node['archiva']['home']}/#{file}" do
     action :delete