+++ spec/spec_helper.rb
@@ -1,14 +1,9 @@
 require 'chefspec'
 require 'chefspec/berkshelf'
-require 'coveralls'
-Coveralls.wear!
 
 RSpec.configure do |config|
   config.formatter = :documentation
   config.color     = true
   config.platform  = 'ubuntu'
-  config.version   = '14.04'
-  config.log_level = :fatal
+  config.version   = '18.04'
 end
-
-at_exit { ChefSpec::Coverage.report! }