+++ metadata.rb
@@ -4,7 +4,7 @@ maintainer_email 'devops@evertrue.com'
 license          'Apache-2.0'
 description      'Installs/Configures Archiva'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '2.0.0'
+version          '3.0.0'
 chef_version     '>= 12.7' if respond_to?(:chef_version)
 source_url       'https://github.com/evertrue/archiva-cookbook'
 issues_url       'https://github.com/evertrue/archiva-cookbook/issues'